Invert Switchbord icons (user report) Separate out the BeanSwitch class
