+++ java/src/jmri/jmrit/display/switchboardEditor/configurexml/BeanSwitchXml.java
@@ -2,6 +2,7 @@ package jmri.jmrit.display.switchboardEditor.configurexml;
 
 import jmri.configurexml.AbstractXmlAdapter;
 import jmri.jmrit.display.switchboardEditor.SwitchboardEditor;
+import jmri.jmrit.display.switchboardEditor.BeanSwitch;
 import org.jdom2.Element;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -25,7 +26,7 @@ public class BeanSwitchXml extends AbstractXmlAdapter {
      */
     @Override
     public Element store(Object o) {
-        SwitchboardEditor.BeanSwitch bs = (SwitchboardEditor.BeanSwitch) o;
+        BeanSwitch bs = (BeanSwitch) o;
         Element element = new Element("beanswitch");
         // include attributes
         element.setAttribute("label", bs.getNameString());
@@ -68,9 +69,10 @@ public class BeanSwitchXml extends AbstractXmlAdapter {
     @Override
     public void load(Element element, Object o) {
         // create the objects
-        //SwitchboardEditor.BeanSwitch bs = (SwitchboardEditor.BeanSwitch) o;
+        //BeanSwitch bs = (BeanSwitch) o;
         log.debug("load xml called");
     }
 
     private static final Logger log = LoggerFactory.getLogger(BeanSwitchXml.class);
+
 }
